requirements.txt
asgiref==3.6.0
attrs==22.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.2
blinker==1.5
bpmappers==1.3
brotlipy==0.7.0
certifi
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
coverage==7.2.2
decorator==5.1.1
Django==4.1.7
exceptiongroup==1.1.1
Faker==18.3.1
Flask==2.2.3
greenlet==2.0.2
guestbook==1.0.0
gunicorn==20.1.0
httpbin==0.7.0
idna==3.4
importlib-metadata==6.0.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
mock==5.0.1
packaging==23.0
Pillow==9.4.0
pluggy==1.0.0
pycparser==2.21
pycrypto==2.6.1
pytest==7.2.2
pytest-cov==4.0.0
python-dateutil==2.8.2
raven==6.10.0
requests==2.28.2
six==1.16.0
soupsieve==2.4
SQLAlchemy==2.0.6
sqlparse==0.4.3
testfixtures==7.1.0
tomli==2.0.1
typing_extensions==4.5.0
unittest-xml-reporting==3.2.0
urllib3==1.26.15
waitress==2.1.2
WebOb==1.8.7
WebTest==3.0.0
Werkzeug==0.16.1
zipp==3.15.0
